+++ esphome/components/esphome/ota/ota_esphome.h
@@ -5,6 +5,7 @@
 #include "esphome/core/defines.h"
 #include "esphome/core/helpers.h"
 #include "esphome/core/preferences.h"
+// #include "esphome/components/ota/ota_backend.h"
 
 namespace esphome {
 namespace ota_esphome {
@@ -15,6 +16,7 @@ enum OTAESPHomeState { OTA_COMPLETED = 0, OTA_STARTED, OTA_IN_PROGRESS, OTA_ERRO
 class OTAESPHomeComponent : public Component {
  public:
   OTAESPHomeComponent();
+  // static std::unique_ptr<ota::OTABackend> make_ota_backend();
 #ifdef USE_OTA_PASSWORD
   void set_auth_password(const std::string &password) { password_ = password; }
 #endif  // USE_OTA_PASSWORD
